+++ src/lib.rs
@@ -15,6 +15,6 @@ pub use read::{DawprojectReadError, DawprojectReader};
 /// Metadata of the `metadata.xml` file.
 pub use repositories::metadata::MetaData;
 /// Project of the `project.xml` file.
-pub use repositories::project::Project;
+pub use repositories::project::ProjectType as Project;
 pub use repositories::*;
 pub use write::{DawprojectWriteError, DawprojectWriter};