+++ src/prelude.rs
@@ -6,6 +6,6 @@ pub use crate::read::{DawprojectReadError, DawprojectReader};
 /// Metadata of the `metadata.xml` file.
 pub use crate::repositories::metadata::MetaData;
 /// Project of the `project.xml` file.
-pub use crate::repositories::project::Project;
+pub use crate::repositories::project::ProjectType as Project;
 pub use crate::repositories::*;
 pub use crate::write::{DawprojectWriteError, DawprojectWriter};